# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API docs for my GO backend's sandbox",
        "title": "Go Rest API docs",
        "contact": {
            "name": "pyncz",
            "url": "http://github.com/pyncz",
            "email": "pyncz.dev@google.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "get": {
                "description": "Reads a paginated list of the tasks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "read tasks",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 12,
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginatedListResults-Task"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HttpError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new task from the provided form",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "create task",
                "parameters": [
                    {
                        "description": "Creation form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TaskCreateForm"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HttpError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HttpError"
                        }
                    }
                }
            }
        },
...
        "/tasks/{id}": {
            "get": {
                "description": "Finds a task by the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "get task by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HttpBadRequestError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HttpError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HttpError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "HttpBadRequestError": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "HttpError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "Task": {
            "description": "Task data",
            "type": "object",
            "required": [
                "id",
                "title"
            ],
            "properties": {
                "completed": {
                    "description": "If the task is completed",
                    "type": "boolean"
                },
                "id": {
                    "description": "ID in the DB collection",
                    "type": "string"
                },
                "tags": {
                    "description": "Task's related tags' IDs",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "description": "Title of the task",
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 3
                }
            }
        },
        "TaskCreateForm": {
            "description": "Task creation model",
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "completed": {
                    "description": "If the task is completed",
                    "type": "boolean"
                },
                "tags": {
                    "description": "Task's related tags' IDs",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "description": "Title of the task",
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 3
                }
            }
        },
...
        "models.PaginatedListResults-Task": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "cursor": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Task"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}